+++ contracts/mocks/land/RefereeMock.sol
@@ -5,4 +5,9 @@ import {MockReferee} from "@gelatonetwork/node-sale-rewards/contracts/MockRefere
 
 contract RefereeMock is MockReferee {
     constructor(address nodeKey) MockReferee(nodeKey) {}
+
+    function claimRewardAllNonSuccessfulAttestations(uint256 _nodeKeyId, uint256 _batchesCount) external {
+        uint256[] memory claimableBatchNumbers = new uint256[](_batchesCount);
+        nodeRewards.claimReward(_nodeKeyId, claimableBatchNumbers);
+    }
 }